Bugfix: make sure the relation is resolved before ->
hitting mergeQueryOptions. There is no ideal solution here.
mergeQueryOptions returns a hash and is not expected to work async.
Hence you need to make sure the things you filter on are present.

You can do this on several layers:
- or in the parent route, but this is brittle that it is not
super-visible
- in the route self, using beforeModel. This is brittle, because it
assumes this hook is called before mergeQueryOptions.

I went for the latter, as this is more visible.