Rearrange saving scheduled jobs
When a scheduled job is created, the scheduled job controller encrypts
the secrets (notified via delta-notifier). If there is too much time
between the job and the rest of the items, the scheduled job controller
is notified too soon and starts querying before all the data is there.
This rearranging helps (but does not solve it completely).